Moved libs files to lib directory
