+++ index.js
@@ -1,11 +1,11 @@
 'use strict';
 
-const AbiBinProvider = require('./libs/AbiBinProvider');
-const EconomySetup = require('./libs/EconomySetup');
-const Contracts = require('./libs/Contracts');
-const StakeHelper = require('./libs/helpers/stake/gateway_composer/StakeHelper');
-const Staker = require('./libs/helpers/stake/gateway_composer/Staker');
-const Facilitator = require('./libs/helpers/stake/gateway_composer/Facilitator');
+const AbiBinProvider = require('./lib/AbiBinProvider');
+const EconomySetup = require('./lib/EconomySetup');
+const Contracts = require('./lib/Contracts');
+const StakeHelper = require('./lib/helpers/stake/gateway_composer/StakeHelper');
+const Staker = require('./lib/helpers/stake/gateway_composer/Staker');
+const Facilitator = require('./lib/helpers/stake/gateway_composer/Facilitator');
 
 module.exports = {
   AbiBinProvider: AbiBinProvider,
