+++ lib/AbiBinProvider.js
