+++ lib/Contracts.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Web3 = require('web3');
-const AbiBinProvider = require('../libs/AbiBinProvider');
+const AbiBinProvider = require('./AbiBinProvider');
 const Mosaic = require('@openstfoundation/mosaic-tbd');
 const Contracts = Mosaic.Contracts;
 let abProvider = new AbiBinProvider();
