+++ lib/helpers/setup/BTHelper.js
