+++ lib/helpers/setup/GCHelper.js
