+++ lib/helpers/stake/gateway_composer/Facilitator.js
