+++ lib/helpers/stake/gateway_composer/StakeHelper.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const AbiBinProvider = require('./../../../AbiBinProvider'),
+const AbiBinProvider = require('../../../AbiBinProvider'),
   gatewayComposerContractName = 'GatewayComposer',
   brandedTokenContractName = 'BrandedToken';
