+++ lib/helpers/stake/gateway_composer/Staker.js
