+++ tests/integration/helpers/stake/Facilitator.js
@@ -11,9 +11,9 @@ const Setup = Package.EconomySetup,
   assert = chai.assert,
   config = require('../../../utils/configReader'),
   Web3WalletHelper = require('../../../utils/Web3WalletHelper'),
-  StakeHelper = require('../../../../libs/helpers/stake/gateway_composer/StakeHelper'),
-  Staker = require('../../../../libs/helpers/stake/gateway_composer/Staker'),
-  Facilitator = require('../../../../libs/helpers/stake/gateway_composer/Facilitator'),
+  StakeHelper = require('../../../../lib/helpers/stake/gateway_composer/StakeHelper'),
+  Staker = require('../../../../lib/helpers/stake/gateway_composer/Staker'),
+  Facilitator = require('../../../../lib/helpers/stake/gateway_composer/Facilitator'),
   MockContractsDeployer = require('../../../utils/MockContractsDeployer'),
   abiBinProvider = MockContractsDeployer.abiBinProvider(),
   BTHelper = Package.EconomySetup.BrandedTokenHelper,
