+++ tests/integration/helpers/stake/StakeHelper.js
@@ -11,7 +11,7 @@ const Setup = Package.EconomySetup,
   assert = chai.assert,
   config = require('../../../utils/configReader'),
   Web3WalletHelper = require('../../../utils/Web3WalletHelper'),
-  StakeHelper = require('../../../../libs/helpers/stake/gateway_composer/StakeHelper'),
+  StakeHelper = require('../../../../lib/helpers/stake/gateway_composer/StakeHelper'),
   MockContractsDeployer = require('../../../utils/MockContractsDeployer'),
   abiBinProvider = MockContractsDeployer.abiBinProvider(),
   BTHelper = Package.EconomySetup.BrandedTokenHelper,
