+++ tests/integration/helpers/stake/Staker.js
@@ -10,8 +10,8 @@ const Setup = Package.EconomySetup,
   assert = chai.assert,
   config = require('../../../utils/configReader'),
   Web3WalletHelper = require('../../../utils/Web3WalletHelper'),
-  StakeHelper = require('../../../../libs/helpers/stake/gateway_composer/StakeHelper'),
-  Staker = require('../../../../libs/helpers/stake/gateway_composer/Staker'),
+  StakeHelper = require('../../../../lib/helpers/stake/gateway_composer/StakeHelper'),
+  Staker = require('../../../../lib/helpers/stake/gateway_composer/Staker'),
   MockContractsDeployer = require('../../../utils/MockContractsDeployer'),
   abiBinProvider = MockContractsDeployer.abiBinProvider(),
   BTHelper = Package.EconomySetup.BrandedTokenHelper,
